apps/youtube-url-saver/requirements.txt
yt-dlp
imageio-ffmpeg
certifi
Pillow
